+++ usb_android/android/src/test/kotlin/dev/yanshouwang/usb_android/USBAndroidPluginPluginTest.kt
@@ -13,7 +13,7 @@ import org.mockito.Mockito
  * you can run them directly from IDEs that support JUnit such as Android Studio.
  */
 
-internal class UsbAndroidPluginTest {
+internal class USBAndroidPluginPluginTest {
   @Test
   fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
     val plugin = UsbAndroidPlugin()