+++ usb_android/lib/src/usb_pigeon.dart
@@ -1,0 +1,1 @@
+export 'usb_pigeon/usb_pigeon.g.dart';