+++ usb_android/lib/usb_android.dart
@@ -1,9 +1,1 @@
-import 'package:usb_platform_interface/usb_platform_interface.dart';
-
-import 'src/my_usb_manager.dart';
-
-class UsbAndroid {
-  static void registerWith() {
-    UsbManager.instance = MyUsbManager();
-  }
-}
+export 'src/usb_android_plugin.dart';