+++ usb_platform_interface/lib/usb_platform_interface.dart
@@ -1,6 +1,8 @@
-export 'src/usb_manager.dart';
 export 'src/usb_accessory.dart';
-export 'src/usb_device.dart';
 export 'src/usb_configuration.dart';
-export 'src/usb_interface.dart';
 export 'src/usb_constants.dart';
+export 'src/usb_device.dart';
+export 'src/usb_device_connection.dart';
+export 'src/usb_interface.dart';
+export 'src/usb_manager.dart';
+export 'src/usb_plugin.dart';